Move prepare (/process) guild handler to its own file
